Updated "glip" dependency, fixed copy flag bug
Previously, copying would break when lots of content was written to the
clipboard, due to a bug in the "glip" package. This commit resolves the
issue by updating "glip", and also disables buffering whenever Opts.Clip
is enabled (since the clipboard should be written to through one single
operation, meaning we require throughput.Dump to write to an external
buffer anyways).